I silenced compiler warnings. I thought I had done this already once

+++ examples/QMP_MILC_test.c
@@ -167,7 +167,7 @@ main (int argc, char** argv)
     it = get_current_time ();
     for (i = 0; i < pargv.loops; i++) {
       value = (int *)QMP_get_memory_pointer(mem);
-      for (k = 0; k < pargv.size/sizeof(int); k++)
+      for (k = 0; k < (int)(pargv.size/sizeof(int)); k++)
 	value[k] = i;
       for (j = 1; j < num_nodes; j++)
 	send_field ((char *)value, pargv.size, j);
@@ -182,7 +182,7 @@ main (int argc, char** argv)
     for (i = 0; i < pargv.loops; i++) {
       value = (int *)QMP_get_memory_pointer(mem);
       get_field ((char *)value, pargv.size, 0);
-      for (k = 0; k < pargv.size/sizeof(int); k++)
+      for (k = 0; k < (int)(pargv.size/sizeof(int)); k++)
 	if (value[k] != i) 
 	  QMP_fprintf (stderr, "Receiving error on loop %d\n", i);
     }
